telegram-bot/requirements.txt
python-telegram-bot==20.7
python-whois==0.9.4
dnspython==2.4.2
builtwith==1.3.6
ipinfo==4.4.2
python-dotenv==1.0.0